roles/redpanda_logging/tests/requirements.txt
pytest==7.4.0
pyyaml==6.0.1
jinja2==3.1.2